build: add Fontconfig & Freetype for text rendering

+++ Generals/Code/Libraries/Source/WWVegas/WWLib/bittype.h
@@ -43,12 +43,12 @@
 
 typedef unsigned char	uint8;
 typedef unsigned short	uint16;
-typedef unsigned long	uint32;
+typedef uint32_t	    uint32;
 typedef unsigned int    uint;
 
 typedef signed char		sint8;
-typedef signed short		sint16;
-typedef signed long		sint32;
+typedef signed short	sint16;
+typedef int32_t		    sint32;
 typedef signed int      sint;
 
 typedef float				float32;
